backend/pyproject.toml
[tool.poetry]
name = "web-server"
version = "2.1.0"
description = "just for test"
authors = ["blacksheep <blacksheep@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
blacksheep = { version = "^2.0.0", extras = ["full"] }
uvicorn = "^0.27.0"
pydantic = "^2.6.3"
httpx = "^0.27.0"
orjson = "^3.9.15"
uvloop = { version = "^0.19.0",  platform = "linux"}
pydantic-settings = "^2.2.1"
aiosqlite = "^0.20.0"
blacksheep-sqlalchemy = "^0.0.3"
alembic = "^1.13.1"
python-crontab = "^3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

